adjustments

+++ src/pages/products-list/products-list.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 import React, { useEffect, useState } from 'react';
 import { ToastContainer, toast } from 'react-toastify';
 import { MdDeleteForever, MdModeEdit } from 'react-icons/md';
@@ -7,6 +8,7 @@ import api from '../../services/api';
 import style from './products-list.module.scss';
 import Loader from '../components/loading';
 import Card from './components/index';
+import emptyImg from '../../static/images/undraw_empty_re_opql.svg';
 
 function RegisterProduct() {
   const history = useHistory();
@@ -61,26 +63,32 @@ function RegisterProduct() {
   };
 
   const handleRemove = async (idRemove, name) => {
-    let existSale = false;
-    sales.map((saleData) => {
-      const productExist = saleData.products.filter(
-        (product) => product.product._id === idRemove
-      );
+    const confirm = window.confirm(
+      'você realmente deseja remover este produto?'
+    );
 
-      if (productExist.length > 0) {
-        existSale = true;
+    if (confirm) {
+      let existSale = false;
+      sales.map((saleData) => {
+        const productExist = saleData.products.filter(
+          (product) => product.product._id === idRemove
+        );
+
+        if (productExist.length > 0) {
+          existSale = true;
+        }
+        return saleData;
+      });
+
+      if (existSale) {
+        toast.error(
+          'Esta ação nao pode ser realizada pois existe uma venda registrada com este produto'
+        );
+      } else {
+        await api.delete(`/products/${idRemove}`);
+        toast.success(`O produto ${name} foi deletado com sucesso!`);
+        findProducts();
       }
-      return saleData;
-    });
-
-    if (existSale) {
-      toast.error(
-        'Esta ação nao pode ser realizada pois existe uma venda registrada com este produto'
-      );
-    } else {
-      await api.delete(`/products/${idRemove}`);
-      toast.success(`O produto ${name} foi deletado com sucesso!`);
-      findProducts();
     }
   };
 
@@ -95,87 +103,103 @@ function RegisterProduct() {
             </div>
           ) : (
             <>
-              <div className={style.Card}>
-                {products?.map((product) => {
-                  const dataFormated = {
-                    ...product,
-                    _id: product.string,
-                    categoryLabel: product.category.label,
-                    description: product.description,
-                    code: product.code,
-                    costSale: product.costSale,
-                    costValue: product.costValue,
-                    amountStock: product.amountStock,
-                    date: product.date,
-                  };
-                  return (
-                    <Card
-                      data={dataFormated}
-                      key={product._id}
-                      handleRemove={handleRemove}
-                      handleEdit={handleEdit}
-                    />
-                  );
-                })}
-              </div>
-              <table className={style.Table}>
-                <tbody>
-                  <tr>
-                    <th>Produto</th>
-                    <th>Descrição</th>
-                    <th>Código</th>
-                    <th>Valor de custo</th>
-                    <th>Valor de venda</th>
-                    <th>Ganho</th>
-                    <th>Estoque</th>
-                    <th>Data</th>
-                    <th>Editar</th>
-                    <th>Excluir</th>
-                  </tr>
+              {products && products.length > 0 ? (
+                <>
+                  <div className={style.Card}>
+                    {products.map((product) => {
+                      const dataFormated = {
+                        ...product,
+                        _id: product.string,
+                        categoryLabel: product.category.label,
+                        description: product.description,
+                        code: product.code,
+                        costSale: product.costSale,
+                        costValue: product.costValue,
+                        amountStock: product.amountStock,
+                        date: product.date,
+                      };
+                      return (
+                        <Card
+                          data={dataFormated}
+                          key={product._id}
+                          handleRemove={handleRemove}
+                          handleEdit={handleEdit}
+                        />
+                      );
+                    })}
+                  </div>
+                  <table className={style.Table}>
+                    <tbody>
+                      <tr>
+                        <th>Produto</th>
+                        <th>Descrição</th>
+                        <th>Código</th>
+                        <th>Valor de custo</th>
+                        <th>Valor de venda</th>
+                        <th>Lucro</th>
+                        <th>Estoque</th>
+                        <th>Data</th>
+                        <th>Editar</th>
+                        <th>Excluir</th>
+                      </tr>
 
-                  {products?.map((product) => (
-                    <tr key={product._id}>
-                      <td>{product.category.label}</td>
-                      <td>{product.description}</td>
-                      <td>{product.code}</td>
-                      <td>{`R$${product.costValue}`}</td>
-                      <td>{`R$${product.costSale}`}</td>
-                      <td>{`R$${product.costSale - product.costValue} `}</td>
-                      <td
-                        style={{
-                          color:
-                            product.amountStock >= 1 ? '#388E3C' : '#D32F2F',
-                        }}
-                      >
-                        {product.amountStock >= 1
-                          ? product.amountStock
-                          : 'Sem estoque'}
-                      </td>
-                      <td>{new Date(product.date).toLocaleDateString()}</td>
-                      <td>
-                        <button
-                          className={style.EditButton}
-                          onClick={() => handleEdit(product)}
-                          type="button"
-                        >
-                          <MdModeEdit />
-                        </button>
-                      </td>
-                      <td>
-                        <button
-                          className={style.DeleteButton}
-                          onClick={() =>
-                            handleRemove(product._id, product.category.label)
-                          }
-                          type="button"
-                        >
-                          <MdDeleteForever />
-                        </button>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
+                      {products.map((product) => (
+                        <tr key={product._id}>
+                          <td>{product.category.label}</td>
+                          <td>{product.description}</td>
+                          <td>{product.code}</td>
+                          <td>{`R$${product.costValue}`}</td>
+                          <td>{`R$${product.costSale}`}</td>
+                          <td>{`R$${
+                            product.costSale - product.costValue
+                          } `}</td>
+                          <td
+                            style={{
+                              color:
+                                product.amountStock >= 1
+                                  ? '#388E3C'
+                                  : '#D32F2F',
+                            }}
+                          >
+                            {product.amountStock >= 1
+                              ? product.amountStock
+                              : 'Sem estoque'}
+                          </td>
+                          <td>{new Date(product.date).toLocaleDateString()}</td>
+                          <td>
+                            <button
+                              className={style.EditButton}
+                              onClick={() => handleEdit(product)}
+                              type="button"
+                            >
+                              <MdModeEdit />
+                            </button>
+                          </td>
+                          <td>
+                            <button
+                              className={style.DeleteButton}
+                              onClick={() =>
+                                handleRemove(
+                                  product._id,
+                                  product.category.label
+                                )
+                              }
+                              type="button"
+                            >
+                              <MdDeleteForever />
+                            </button>
+                          </td>
+                        </tr>
+                      ))}
+                    </tbody>
+                  </table>
+                </>
+              ) : (
+                <div className={style.Emptyimg}>
+                  <h1>Não há produtos cadastrados</h1>
+                  <img alt="Não existem produtos" src={emptyImg} />
+                </div>
+              )}
             </>
           )}
         </div>
