+++ src/pages/sales-list/sales-list.jsx
@@ -12,6 +12,7 @@ import Layout from '../layouts';
 import api from '../../services/api';
 import Loader from '../components/loading';
 import Card from './components/card';
+import emptyImg from '../../static/images/undraw_empty_re_opql.svg';
 
 import style from './sales-list.module.scss';
 
@@ -90,9 +91,13 @@ function SalesList() {
   };
 
   const handleRemove = async (idRemove) => {
-    await api.delete(`/sales/${idRemove}`, { isActive: false });
-    toast.success(`A venda foi deletada com sucesso!`);
-    findSales();
+    const confirm = window.confirm('você realmente deseja remover esta venda?');
+
+    if (confirm) {
+      await api.delete(`/sales/${idRemove}`, { isActive: false });
+      toast.success(`A venda foi deletada com sucesso!`);
+      findSales();
+    }
   };
 
   const onButtonClick = async (saleId, index, handleModal) => {
@@ -131,118 +136,124 @@ function SalesList() {
             </div>
           ) : (
             <>
-              <div className={style.SalesCard}>
-                {sales &&
-                  sales.length > 0 &&
-                  sales.map((sale, index) => {
-                    const saleDataFormated = {
-                      ...sale,
-                      _id: sale._id,
-                      itsPaid: sale.itsPaid,
-                      clientName: sale.client?.name,
-                      date: sale.date,
-                      totalValue: sale.totalValue,
-                      totalValuePaid: sale.totalValuePaid,
-                    };
-                    return (
-                      <Card
-                        data={saleDataFormated}
-                        onButtonClick={onButtonClick}
-                        index={index}
-                        inputPaymentRef={inputPaymentRef}
-                        inputDataRef={inputDataRef}
-                        key={sale._id}
-                        handleEdit={handleEdit}
-                        handleRemove={handleRemove}
-                      />
-                    );
-                  })}
-              </div>
-              <table className={style.Table}>
-                <tbody>
-                  <tr>
-                    <th>Status</th>
-                    <th>Data</th>
-                    <th>Cliente</th>
-                    <th>Valor total</th>
-                    <th>Valor Pago</th>
-                    <th>Adicionar pagamento</th>
-                    <th>Lucro na venda</th>
-                    <th>Valor de custo</th>
-                    <th>Editar</th>
-                    <th>Exluir</th>
-                  </tr>
-                  {sales.length > 0 &&
-                    sales.map((sale, i) => (
-                      <tr key={sale._id}>
-                        <td>
-                          <div>
-                            {sale.itsPaid ? (
-                              <>
-                                <MdDoneAll style={{ color: '#388E3C' }} />
-                                Pago
-                              </>
-                            ) : (
-                              <>
-                                <MdRemoveDone style={{ color: '#D32F2F' }} />
-                                Não pago
-                              </>
-                            )}
-                          </div>
-                        </td>
-                        <td>{new Date(sale.date).toLocaleDateString()}</td>
-                        <td>{sale.client?.name}</td>
-                        <td>{`R$${sale?.totalValue}`}</td>
-                        <td>{`R$${sale?.totalValuePaid}`}</td>
-                        <td>
-                          <div className={style.AddPayment}>
-                            <input
-                              type="number"
-                              min="1"
-                              ref={(el) => {
-                                inputPaymentRef.current[i] = el;
-                                return el;
-                              }}
-                            />
-                            <input
-                              type="date"
-                              ref={(el) => {
-                                inputDataRef.current[i] = el;
-                                return el;
-                              }}
-                            />
-                            <button
-                              type="button"
-                              onClick={() => onButtonClick(sale._id, i)}
-                            >
-                              <MdOutlineAddCircle
-                                style={{ color: '#388E3C' }}
-                              />
-                            </button>
-                          </div>
-                        </td>
-                        <td>{`R$${sale?.lucre}`}</td>
-                        <td>{`R$${sale?.costValue}`}</td>
-                        <td>
-                          <button
-                            onClick={() => handleEdit(sale)}
-                            type="button"
-                          >
-                            <MdModeEdit className={style.EditButton} />
-                          </button>
-                        </td>
-                        <td>
-                          <button
-                            onClick={() => handleRemove(sale._id)}
-                            type="button"
-                          >
-                            <MdDeleteForever className={style.DeleteButton} />
-                          </button>
-                        </td>
+              {sales && sales.length > 0 ? (
+                <>
+                  <div className={style.SalesCard}>
+                    {sales.map((sale, index) => {
+                      const saleDataFormated = {
+                        ...sale,
+                        _id: sale._id,
+                        itsPaid: sale.itsPaid,
+                        clientName: sale.client?.name,
+                        date: sale.date,
+                        totalValue: sale.totalValue,
+                        totalValuePaid: sale.totalValuePaid,
+                      };
+                      return (
+                        <Card
+                          data={saleDataFormated}
+                          onButtonClick={onButtonClick}
+                          index={index}
+                          inputPaymentRef={inputPaymentRef}
+                          inputDataRef={inputDataRef}
+                          key={sale._id}
+                          handleEdit={handleEdit}
+                          handleRemove={handleRemove}
+                        />
+                      );
+                    })}
+                  </div>
+                  <table className={style.Table}>
+                    <tbody>
+                      <tr>
+                        <th>Status</th>
+                        <th>Data</th>
+                        <th>Cliente</th>
+                        <th>Valor total</th>
+                        <th>Valor Pago</th>
+                        <th>Adicionar pagamento</th>
+                        <th>Lucro na venda</th>
+                        <th>Valor de custo</th>
+                        <th>Editar</th>
+                        <th>Exluir</th>
                       </tr>
-                    ))}
-                </tbody>
-              </table>
+                      {sales.map((sale, i) => (
+                        <tr key={sale._id}>
+                          <td>
+                            <div>
+                              {sale.itsPaid ? (
+                                <>
+                                  <MdDoneAll style={{ color: '#388E3C' }} />
+                                  Pago
+                                </>
+                              ) : (
+                                <>
+                                  <MdRemoveDone style={{ color: '#D32F2F' }} />
+                                  Não pago
+                                </>
+                              )}
+                            </div>
+                          </td>
+                          <td>{new Date(sale.date).toLocaleDateString()}</td>
+                          <td>{sale.client?.name}</td>
+                          <td>{`R$${sale?.totalValue}`}</td>
+                          <td>{`R$${sale?.totalValuePaid}`}</td>
+                          <td>
+                            <div className={style.AddPayment}>
+                              <input
+                                type="number"
+                                min="1"
+                                ref={(el) => {
+                                  inputPaymentRef.current[i] = el;
+                                  return el;
+                                }}
+                              />
+                              <input
+                                type="date"
+                                ref={(el) => {
+                                  inputDataRef.current[i] = el;
+                                  return el;
+                                }}
+                              />
+                              <button
+                                type="button"
+                                onClick={() => onButtonClick(sale._id, i)}
+                              >
+                                <MdOutlineAddCircle
+                                  style={{ color: '#388E3C' }}
+                                />
+                              </button>
+                            </div>
+                          </td>
+                          <td>{`R$${sale?.lucre}`}</td>
+                          <td>{`R$${sale?.costValue}`}</td>
+                          <td>
+                            <button
+                              onClick={() => handleEdit(sale)}
+                              type="button"
+                            >
+                              <MdModeEdit className={style.EditButton} />
+                            </button>
+                          </td>
+                          <td>
+                            <button
+                              onClick={() => handleRemove(sale._id)}
+                              type="button"
+                            >
+                              <MdDeleteForever className={style.DeleteButton} />
+                            </button>
+                          </td>
+                        </tr>
+                      ))}
+                    </tbody>
+                  </table>
+                </>
+              ) : (
+                <div className={style.Emptyimg}>
+                  <h1>Não há vendas cadastrados</h1>
+                  <img alt="Não existem vendas" src={emptyImg} />
+                </div>
+              )}
             </>
           )}
         </div>
